requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
joblib==0.14.1
nltk==3.4.5
numpy==1.18.3
onesignal-notifications==0.2.6
pytz==2019.3
requests==2.23.0
scikit-learn==0.21.3
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
